refactor: errorDirectoryAdd && StatusCode modify
