+++ src/errors/party.error.js
@@ -41,3 +41,16 @@ export class InvalidPartyPasswordError extends Error {
     this.statusCode = 401;
   }
 }
+
+
+//파티 생성 시 존재하는 이름일 경우
+export class ExsistsPartyNameError extends Error {
+  errorCode = 'P006';
+
+  constructor(reason, data) {
+    super(reason);
+    this.reason = reason;
+    this.data = data;
+    this.statusCode = 400;
+  }
+}
